修改logout的问题

+++ src/main/java/com/parknshop/controller/UserCenterController.java
@@ -3,6 +3,7 @@ package com.parknshop.controller;
 import com.parknshop.entity.CollectionEntity;
 import com.parknshop.entity.CollectshopEntity;
 import com.parknshop.entity.UserEntity;
+import com.parknshop.service.IUserService;
 import com.parknshop.service.baseImpl.IDefineString;
 import com.parknshop.service.customerService.ICustomerService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,12 +27,14 @@ public class UserCenterController {
     @Autowired
     private ICustomerService customerService;
 
+    @Autowired
+    private IUserService mUserService;
 
     @RequestMapping(value="/listUserInfo",method = RequestMethod.GET)
     public String  listUserInfo(Model model, HttpSession session)
     {
         int userId=getUserId(session);
-        if(userId<1){
+        if (userId<0) {
             return "redirect:customer/login";
         }else{
             UserEntity userEntity=customerService.getCustomerById(new Integer(userId));
@@ -50,7 +53,8 @@ public class UserCenterController {
     public String  goPassword(){
         return "customer/password_edit.jsp";
     }
-//此处需要前端验证两次密码一致性，且应用Ajax判断密码的正确性，直至输对密码为止
+
+
     @RequestMapping(value="/changePassword",method = RequestMethod.POST)
     public String  changePassword(@RequestParam String  password, @RequestParam String  pass1, Model model,HttpSession session)
     {
@@ -75,12 +79,12 @@ public class UserCenterController {
     public String listCollect(@RequestParam int requestPage, Model  model,HttpSession session)
     {
         int userId=getUserId(session);
-        if(userId<1){
+        if (userId<0) {
             return "redirect:customer/login";
         }else{
             int size = customerService.querySize(userId);
-            double d = size / 4;
-            int  sina =(int) Math.floor(d)+1;
+            double d =  (size) / 4+(size%4==0?0:1);
+            int  sina =(int) Math.ceil(d);
             List<CollectionEntity> collectionEntities = customerService.queryAllCollect(new Integer(userId), requestPage);
             model.addAttribute("Collects", collectionEntities);
             model.addAttribute("currentPage", requestPage);
@@ -95,12 +99,12 @@ public class UserCenterController {
     public String listCollectShop(@RequestParam int requestPage, Model  model,HttpSession session)
     {
         int userId=getUserId(session);
-        if(userId<1){
+        if (userId<0) {
             return "redirect:customer/login";
         }else{
             int size = customerService.queryShopsize(new Integer(userId));
-            double d = size / 4;
-            int  sina =(int) Math.floor(d)+1;
+            double d =  (size) / 4+(size%4==0?0:1);
+            int  sina =(int) Math.ceil(d);
             List<CollectshopEntity> collectionEntities = customerService.queryAllShop(new Integer(userId), requestPage);
             model.addAttribute("Collects", collectionEntities);
             model.addAttribute("currentPage", requestPage);
@@ -113,10 +117,11 @@ public class UserCenterController {
     @RequestMapping(value="/insertCollectShop",method = RequestMethod.GET)
     public String  insertCollectShop(@RequestParam int  shopId,HttpSession session)
     {
-        int userId=getUserId(session);
-        if(userId<1){
+
+        if (getUserId(session)<0) {
             return "redirect:customer/login";
         }else{
+            int userId=getUserId(session);
             customerService.insertShop(shopId,userId);
             return "redirect:listCollectShop?requestPage=1";
         }
@@ -127,7 +132,7 @@ public class UserCenterController {
     public String  insertCollect(@RequestParam int  goodsId,HttpSession session)
     {
         int userId=getUserId(session);
-        if(userId<1){
+        if (userId<0) {
             return "redirect:customer/login";
         }else{
             customerService.insertCollect(goodsId, userId);
@@ -136,21 +141,32 @@ public class UserCenterController {
     }
 
     @RequestMapping(value="/removeCollect",method = RequestMethod.GET)
-    public String removeCollect(@RequestParam int collectionId)
+    public String removeCollect(@RequestParam int collectionId,HttpSession session)
     {
-        customerService.removeCollect(new Integer(collectionId));
+        if (getUserId(session)<0) {
+            return "redirect:customer/login";
+        }
+        else{
+            customerService.removeCollect(new Integer(collectionId));
             return "redirect:listCollect?requestPage=1";
+        }
     }
 
     @RequestMapping(value="/removeCollectShop",method = RequestMethod.GET)
-    public String removeCollectShop(@RequestParam int shopId)
+    public String removeCollectShop(@RequestParam int shopId,HttpSession session)
     {
-        customerService.removeShop(new Integer(shopId));
-        return "redirect:listCollectShop?requestPage=1";
+        if (getUserId(session)<0) {
+            return "redirect:customer/login";
+        }
+        else {
+            customerService.removeShop(new Integer(shopId));
+            return "redirect:listCollectShop?requestPage=1";
+        }
     }
     @RequestMapping(value="/jumpPage",method = RequestMethod.POST)
     public String jumpPage(@RequestParam int jump,@RequestParam int ty, Model model)
     {
+
         if(ty==1) {
             model.addAttribute("gdhds", jump);
             return "redirect:listCollect?requestPage={gdhds}";
